Final Touches

+++ CS401 Assignments/CS401 Assignment 4/src/UserArtistGraph.java
@@ -25,7 +25,7 @@ class UserArtistGraph
         {
             int v = in.readInt();
             int w = in.readInt();
-            int weight = in.readInt();
+            double weight = in.readDouble();
             DirectedEdge directedEdge = new DirectedEdge(v, w, weight);
             edgeWeightedDigraph.addEdge(directedEdge);
         }
